+++ typechain-types/factories/contracts/interfaces/IMasterPool.sol/IMasterPool__factory.ts
@@ -35,14 +35,14 @@ const _abi = [
         type: "uint32",
       },
       {
-        internalType: "uint64",
+        internalType: "uint48",
         name: "_startTime",
-        type: "uint64",
+        type: "uint48",
       },
       {
-        internalType: "uint64",
+        internalType: "uint48",
         name: "_endTime",
-        type: "uint64",
+        type: "uint48",
       },
       {
         internalType: "bool",
@@ -92,9 +92,9 @@ const _abi = [
         type: "uint256",
       },
       {
-        internalType: "uint64",
+        internalType: "uint48",
         name: "_deadline",
-        type: "uint64",
+        type: "uint48",
       },
       {
         internalType: "bytes",
@@ -115,14 +115,14 @@ const _abi = [
         type: "uint16",
       },
       {
-        internalType: "uint64",
+        internalType: "uint48",
         name: "_fromTime",
-        type: "uint64",
+        type: "uint48",
       },
       {
-        internalType: "uint64",
+        internalType: "uint48",
         name: "_toTime",
-        type: "uint64",
+        type: "uint48",
       },
     ],
     name: "getMultiplier",
@@ -144,14 +144,14 @@ const _abi = [
         type: "uint16",
       },
       {
-        internalType: "uint64",
+        internalType: "uint48",
         name: "_fromTime",
-        type: "uint64",
+        type: "uint48",
       },
       {
-        internalType: "uint64",
+        internalType: "uint48",
         name: "_toTime",
-        type: "uint64",
+        type: "uint48",
       },
     ],
     name: "getPoolRewardsByTime",
@@ -208,9 +208,9 @@ const _abi = [
     name: "halvingInterval",
     outputs: [
       {
-        internalType: "uint64",
+        internalType: "uint48",
         name: "",
-        type: "uint64",
+        type: "uint48",
       },
     ],
     stateMutability: "view",
@@ -295,14 +295,14 @@ const _abi = [
         type: "uint256",
       },
       {
-        internalType: "uint64",
+        internalType: "uint48",
         name: "_startTime",
-        type: "uint64",
+        type: "uint48",
       },
       {
-        internalType: "uint64",
+        internalType: "uint48",
         name: "_endTime",
-        type: "uint64",
+        type: "uint48",
       },
       {
         internalType: "bool",
@@ -353,19 +353,19 @@ const _abi = [
         type: "uint16",
       },
       {
-        internalType: "uint64",
+        internalType: "uint48",
         name: "_halvingInterval",
-        type: "uint64",
+        type: "uint48",
       },
       {
-        internalType: "uint64",
+        internalType: "uint48",
         name: "_startTime",
-        type: "uint64",
+        type: "uint48",
       },
       {
-        internalType: "uint64",
+        internalType: "uint48",
         name: "_endTime",
-        type: "uint64",
+        type: "uint48",
       },
       {
         internalType: "bool",
@@ -488,19 +488,19 @@ const _abi = [
             type: "uint32",
           },
           {
-            internalType: "uint64",
+            internalType: "uint48",
             name: "startTime",
-            type: "uint64",
+            type: "uint48",
           },
           {
-            internalType: "uint64",
+            internalType: "uint48",
             name: "endTime",
-            type: "uint64",
+            type: "uint48",
           },
           {
-            internalType: "uint64",
+            internalType: "uint48",
             name: "lastRewardTime",
-            type: "uint64",
+            type: "uint48",
           },
           {
             internalType: "uint256",
@@ -667,14 +667,14 @@ const _abi = [
         type: "uint32",
       },
       {
-        internalType: "uint64",
+        internalType: "uint48",
         name: "_startTime",
-        type: "uint64",
+        type: "uint48",
       },
       {
-        internalType: "uint64",
+        internalType: "uint48",
         name: "_endTime",
-        type: "uint64",
+        type: "uint48",
       },
       {
         internalType: "bool",
@@ -690,9 +690,9 @@ const _abi = [
   {
     inputs: [
       {
-        internalType: "uint64",
+        internalType: "uint48",
         name: "_halvingInterval",
-        type: "uint64",
+        type: "uint48",
       },
     ],
     name: "setHalvingInterval",
